+++ InventoryTools/EquipmentSuggest/EquipmentSuggestGrid.cs
@@ -6,10 +6,13 @@ using AllaganLib.GameSheets.Sheets;
 using AllaganLib.GameSheets.Sheets.Rows;
 using AllaganLib.Interface.Grid;
 using AllaganLib.Shared.Extensions;
+using CriticalCommonLib.Models;
+using CriticalCommonLib.Services;
 using CriticalCommonLib.Services.Mediator;
 using DalaMock.Host.Mediator;
 using Dalamud.Bindings.ImGui;
 using InventoryTools.Logic;
+using InventoryTools.Logic.Editors;
 
 namespace InventoryTools.EquipmentSuggest;