+++ InventoryTools/Logic/FilterConfiguration.cs
@@ -22,15 +22,15 @@ namespace InventoryTools.Logic
     {
         [JsonIgnore] //Stops the object copy from erasing the field
         private readonly CraftList.Factory _craftListFactory;
-        private List<(ulong, InventoryCategory)> _destinationInventories = new();
+        private List<(ulong, InventoryCategory)> _destinationInventories = [];
         private bool _displayInTabs = true;
         private bool? _duplicatesOnly;
-        private List<uint> _equipSlotCategoryId = new();
+        private List<uint> _equipSlotCategoryId = [];
         private bool? _isCollectible;
         private bool? _isHq;
-        private List<uint> _itemSearchCategoryId = new();
-        private List<uint> _itemSortCategoryId = new();
-        private List<uint> _itemUiCategoryId = new();
+        private List<uint> _itemSearchCategoryId = [];
+        private List<uint> _itemSortCategoryId = [];
+        private List<uint> _itemUiCategoryId = [];
         private Dictionary<string, bool>? _booleanFilters = new();
         private Dictionary<string, string>? _stringFilters = new();
         private Dictionary<string, int>? _integerFilters = new();
@@ -74,7 +74,7 @@ namespace InventoryTools.Logic
         private bool _openAsWindow = false;
         private bool? _canBeBought;
         private bool? _isAvailableAtTimedNode;
-        private List<(ulong, InventoryCategory)> _sourceInventories = new();
+        private List<(ulong, InventoryCategory)> _sourceInventories = [];
         private FilterType _filterType;
         private Vector4? _highlightColor;
         private Vector4? _tabHighlightColor;
@@ -1009,7 +1009,7 @@ namespace InventoryTools.Logic
         {
             if (_columns == null)
             {
-                _columns = new List<ColumnConfiguration>();
+                _columns = [];
             }
             _columns.Add(column);
             if (notify)
@@ -1023,7 +1023,7 @@ namespace InventoryTools.Logic
         {
             if (_craftColumns == null)
             {
-                _craftColumns = new List<ColumnConfiguration>();
+                _craftColumns = [];
             }
             _craftColumns.Add(craftColumn);
             if (notify)
@@ -1037,7 +1037,7 @@ namespace InventoryTools.Logic
         {
             if (_curatedItems == null)
             {
-                _curatedItems = new();
+                _curatedItems = [];
             }
             _curatedItems.Add(curatedItem);
             ConfigurationDirty = true;
@@ -1047,7 +1047,7 @@ namespace InventoryTools.Logic
         {
             if (_curatedItems == null)
             {
-                _curatedItems = new();
+                _curatedItems = [];
             }
             _curatedItems.Remove(curatedItem);
             ConfigurationDirty = true;
@@ -1057,7 +1057,7 @@ namespace InventoryTools.Logic
         {
             if (_curatedItems == null)
             {
-                _curatedItems = new();
+                _curatedItems = [];
             }
             _curatedItems.Clear();
             ConfigurationDirty = true;
@@ -1120,9 +1120,9 @@ namespace InventoryTools.Logic
 
         public bool? GetBooleanFilter(string key)
         {
-            if (BooleanFilters.ContainsKey(key))
+            if (BooleanFilters.TryGetValue(key, out var value))
             {
-                return BooleanFilters[key];
+                return value;
             }
 
             return null;
@@ -1130,9 +1130,9 @@ namespace InventoryTools.Logic
 
         public Vector4? GetColorFilter(string key)
         {
-            if (ColorFilters.ContainsKey(key))
+            if (ColorFilters.TryGetValue(key, out var value))
             {
-                return ColorFilters[key];
+                return value;
             }
 
             return null;
@@ -1140,19 +1140,14 @@ namespace InventoryTools.Logic
 
         public string GetStringFilter(string key)
         {
-            if (StringFilters.ContainsKey(key))
-            {
-                return StringFilters[key];
-            }
-
-            return "";
+            return StringFilters.GetValueOrDefault(key, "");
         }
 
         public int? GetIntegerFilter(string key)
         {
-            if (IntegerFilters.ContainsKey(key))
+            if (IntegerFilters.TryGetValue(key, out var value))
             {
-                return (int?)IntegerFilters[key];
+                return value;
             }
 
             return null;
@@ -1160,9 +1155,9 @@ namespace InventoryTools.Logic
 
         public int? GetDecimalFilter(string key)
         {
-            if (DecimalFilters.ContainsKey(key))
+            if (DecimalFilters.TryGetValue(key, out var value))
             {
-                return (int?)DecimalFilters[key];
+                return (int?)value;
             }
 
             return null;
@@ -1170,19 +1165,19 @@ namespace InventoryTools.Logic
 
         public List<uint> GetUintChoiceFilter(string key)
         {
-            if (UintChoiceFilters.ContainsKey(key))
+            if (UintChoiceFilters.TryGetValue(key, out var value))
             {
-                return UintChoiceFilters[key];
+                return value;
             }
 
-            return new List<uint>();
+            return [];
         }
 
         public uint? GetUintFilter(string key)
         {
-            if (UintFilters.ContainsKey(key))
+            if (UintFilters.TryGetValue(key, out var value))
             {
-                return UintFilters[key];
+                return value;
             }
 
             return null;
@@ -1190,37 +1185,37 @@ namespace InventoryTools.Logic
 
         public List<ulong> GetUlongChoiceFilter(string key)
         {
-            if (UlongChoiceFilters.ContainsKey(key))
+            if (UlongChoiceFilters.TryGetValue(key, out var value))
             {
-                return UlongChoiceFilters[key];
+                return value;
             }
 
-            return new List<ulong>();
+            return [];
         }
 
         public List<string> GetStringChoiceFilter(string key)
         {
-            if (StringChoiceFilters.ContainsKey(key))
+            if (StringChoiceFilters.TryGetValue(key, out var value))
             {
-                return StringChoiceFilters[key];
+                return value;
             }
 
-            return new List<string>();
+            return [];
         }
 
         public void GetFilter(string key, out List<CharacterSearchScope>? value)
         {
-            value = CharacterSearchScopes.ContainsKey(key) ? CharacterSearchScopes[key] : null;
+            CharacterSearchScopes.TryGetValue(key, out value);
         }
 
         public void GetFilter(string key, out List<InventorySearchScope>? value)
         {
-            value = InventorySearchScopes.ContainsKey(key) ? InventorySearchScopes[key] : null;
+            InventorySearchScopes.TryGetValue(key, out value);
         }
 
         public void UpdateBooleanFilter(string key, bool value)
         {
-            if (BooleanFilters.ContainsKey(key) && BooleanFilters[key] == value)
+            if (BooleanFilters.TryGetValue(key, out var currentValue) && currentValue == value)
             {
                 return;
             }
@@ -1232,7 +1227,7 @@ namespace InventoryTools.Logic
 
         public void UpdateColorFilter(string key, Vector4 value)
         {
-            if (ColorFilters.ContainsKey(key) && ColorFilters[key] == value)
+            if (ColorFilters.TryGetValue(key, out var currentValue) && currentValue == value)
             {
                 return;
             }
@@ -1244,9 +1239,8 @@ namespace InventoryTools.Logic
 
         public void RemoveBooleanFilter(string key)
         {
-            if (BooleanFilters.ContainsKey(key))
+            if (BooleanFilters.Remove(key))
             {
-                BooleanFilters.Remove(key);
                 NeedsRefresh = true;
                 ConfigurationDirty = true;
             }
@@ -1254,9 +1248,8 @@ namespace InventoryTools.Logic
 
         public void RemoveColorFilter(string key)
         {
-            if (ColorFilters.ContainsKey(key))
+            if (ColorFilters.Remove(key))
             {
-                ColorFilters.Remove(key);
                 NeedsRefresh = true;
                 ConfigurationDirty = true;
             }
@@ -1264,7 +1257,7 @@ namespace InventoryTools.Logic
 
         public void UpdateStringFilter(string key, string value)
         {
-            if (StringFilters.ContainsKey(key) && StringFilters[key] == value)
+            if (StringFilters.TryGetValue(key, out var currentValue) && currentValue == value)
             {
                 return;
             }
@@ -1276,15 +1269,15 @@ namespace InventoryTools.Logic
 
         public void UpdateIntegerFilter(string key, int? value)
         {
-            if (IntegerFilters.ContainsKey(key) && IntegerFilters[key] == value)
+            if (IntegerFilters.TryGetValue(key, out var currentValue) && currentValue == value)
             {
                 return;
             }
-            if (IntegerFilters.ContainsKey(key) && value == null)
+            if (value == null)
             {
                 IntegerFilters.Remove(key);
             }
-            else if (value != null)
+            else
             {
                 IntegerFilters[key] = value.Value;
             }
@@ -1295,15 +1288,15 @@ namespace InventoryTools.Logic
 
         public void UpdateDecimalFilter(string key, decimal? value)
         {
-            if (DecimalFilters.ContainsKey(key) && DecimalFilters[key] == value)
+            if (DecimalFilters.TryGetValue(key, out var currentValue) && currentValue == value)
             {
                 return;
             }
-            if (DecimalFilters.ContainsKey(key) && value == null)
+            if (value == null)
             {
                 DecimalFilters.Remove(key);
             }
-            else if (value != null)
+            else
             {
                 DecimalFilters[key] = value.Value;
             }
@@ -1321,11 +1314,15 @@ namespace InventoryTools.Logic
 
         public void UpdateUintFilter(string key, uint? value)
         {
-            if (value == null && UintFilters.ContainsKey(key))
+            if (UintFilters.TryGetValue(key, out var currentValue) && currentValue == value)
+            {
+                return;
+            }
+            if (value == null)
             {
                 UintFilters.Remove(key);
             }
-            else if(value != null)
+            else
             {
                 UintFilters[key] = value.Value;
             }
@@ -1352,13 +1349,13 @@ namespace InventoryTools.Logic
             if (value == null)
             {
                 CharacterSearchScopes.Remove(key);
-                ConfigurationDirty = true;
             }
             else
             {
                 CharacterSearchScopes[key] = value;
-                ConfigurationDirty = true;
             }
+
+            ConfigurationDirty = true;
         }
 
         public void SetFilter(string key, List<InventorySearchScope>? value)
@@ -1366,13 +1363,13 @@ namespace InventoryTools.Logic
             if (value == null)
             {
                 InventorySearchScopes.Remove(key);
-                ConfigurationDirty = true;
             }
             else
             {
                 InventorySearchScopes[key] = value;
-                ConfigurationDirty = true;
             }
+
+            ConfigurationDirty = true;
         }
 
         public Dictionary<string, bool> BooleanFilters