+++ InventoryTools/Logic/Filters/GenericHasUseCategoryFilter.cs
@@ -14,6 +14,7 @@ public class GenericHasUseCategoryFilter : BooleanFilter, IGenericFilter
     private readonly ItemInfoRenderCategory _renderCategory;
     private readonly ItemInfoRenderService _infoRenderService;
     private ItemInfoType[]? _useCats;
+    private readonly string _key;
 
     public override int LabelSize { get; set; } = 250;
 
@@ -22,13 +23,11 @@ public class GenericHasUseCategoryFilter : BooleanFilter, IGenericFilter
     {
         _renderCategory = renderCategory;
         _infoRenderService = infoRenderService;
+        _key = "HasUseCat" + (uint)_renderCategory;
     }
 
     public override string Key {
-        get
-        {
-            return "HasUseCat" + (uint)_renderCategory;
-        }
+        get => _key;
         set
         {